+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 - Remove duplicate definition of LocalSearch (#638)
 - Move priority check to Job constructor instead of input parser (#628)
+- Wrong index values without custom matrix (#683)
 - Assignments instead of equality checks in some plan mode assertions (#701)
 - Initialization of single-entry matrix (#699)